refactor upload page into services
