+++ services/upload/__init__.py
@@ -1,0 +1,9 @@
+from .processing import UploadProcessingService
+from .ai import AISuggestionService
+from .modal import ModalService
+
+__all__ = [
+    "UploadProcessingService",
+    "AISuggestionService",
+    "ModalService",
+]
