+++ tests/test_process_uploaded_files.py
@@ -2,6 +2,7 @@ import base64
 import pandas as pd
 
 from pages.file_upload import Callbacks, _uploaded_data_store
+from services.upload import UploadProcessingService
 
 
 def test_multi_part_upload_row_count():
@@ -17,6 +18,7 @@ def test_multi_part_upload_row_count():
     part2 = prefix + b64[mid:]
 
     cb = Callbacks()
+    cb.processing = UploadProcessingService(_uploaded_data_store)
     res = cb.process_uploaded_files([part1, part2], ["sample.csv", "sample.csv"])
     info = res[2]
     assert info["sample.csv"]["rows"] == len(df)
